online_notes_backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aioredis==1.3.1
aiosignal==1.3.1
amqp==2.6.1
apispec==6.8.1
asgiref==3.8.1
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
awscli==1.37.12
backend==0.2.4.1
billiard==3.6.4.0
black==24.8.0
bleach==3.3.0
boto3==1.36.12
botocore==1.36.12
certifi==2020.12.5
cffi==1.14.5
channels==4.2.2
channels_redis==4.2.1
chardet==4.0.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.2.0
dataclasses==0.6
deepdiff==8.0.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.2.1
django-admin-tools==0.9.1
django-allauth==0.55.2
django-autoslug==1.9.8
django-cleanup==8.1.0
django-cors-headers==3.7.0
django-crontab==0.7.0
django-environ==0.4.5
django-filter==2.4.0
django-fobi==0.17.1
django-formtools==2.2
django-grappelli==2.14.3
django-jazzmin==2.4.4
django-nine==0.2.4
django-nonefield==0.4
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djongo==1.2.36
dnspython==2.6.1
docutils==0.16
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.2.1
drf-yasg==1.21.8
easy-thumbnails==2.10
elastic-transport==8.15.0
elasticsearch==8.15.0
Faker==35.2.0
filelock==3.18.0
frozenlist==1.4.1
gunicorn==23.0.0
h11==0.16.0
hiredis==3.1.1
hyperlink==21.0.0
idna==2.10
incremental==24.7.2
inflection==0.5.1
iniconfig==1.1.1
iso8601==0.1.14
Jinja2==3.1.5
jmespath==0.10.0
jsonpickle==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==4.6.11
lxml==5.3.0
MarkupSafe==3.0.2
mimesis==18.0.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.2
oauthlib==3.1.0
orderly-set==5.2.2
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pdf2image==1.16.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prompt-toolkit==3.0.19
psycopg2==2.9.9
psycopg2-binary==2.8.6
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.20
PyJWT==2.9.0
pymongo==4.8.0
pyOpenSSL==24.2.1
pyotp==2.6.0
pyparsing==2.4.7
pytest==8.3.4
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python-decouple==3.4
python-docx==1.1.2
python3-openid==3.2.0
pytz==2021.1
pyxirr==0.10.6
PyYAML==6.0.2
QuantLib==1.37
razorpay==1.4.2
redis==5.2.1
referencing==0.36.2
requests==2.25.1
requests-oauthlib==1.3.0
rpds-py==0.22.3
rsa==4.7.2
s3transfer==0.11.2
scipy==1.14.1
sentry-sdk==2.13.0
service-identity==24.1.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
swagger_ui_bundle==1.1.0
toml==0.10.2
tomli==2.0.1
twilio==9.3.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.2.0
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.2
vine==1.3.0
virtualenv==20.31.2
vishap==0.1.5
waitress==3.0.2
wcwidth==0.2.5
webencodings==0.5.1
yarl==1.11.1
zope.interface==7.0.3
